Feat: Add Repository of FavoriteMachine, Machine, Question, Vote

+++ BackEnd/ggong/src/main/java/com/a304/ggong/entity/Question.java
@@ -20,7 +20,7 @@ public class Question {
 	@Id
 	@GeneratedValue
 	@Column(name = "question_ID")
-	private int questionID;
+	private Long questionID;
 
 	@Column(name = "content")
 	private String content;
